updated backend files

+++ controllers/jobController.js
@@ -23,7 +23,6 @@ export const createJob = async (req, res) => {
       location,
       salary,
       description,
-
       recruiter: req.user.id,
     });
 
@@ -31,6 +30,7 @@ export const createJob = async (req, res) => {
       message: "Job created successfully",
       job,
     });
+
   } catch (error) {
     res.status(500).json({
       message: error.message,
@@ -40,13 +40,16 @@ export const createJob = async (req, res) => {
 
 export const getAllJobs = async (req, res) => {
   try {
-    const jobs = await Job.find().populate("recruiter", "name email role");
+
+    const jobs = await Job.find()
+      .populate("recruiter", "name email role");
 
     res.status(200).json({
       message: "Jobs fetched successfully",
       count: jobs.length,
       jobs,
     });
+
   } catch (error) {
     res.status(500).json({
       message: error.message,
@@ -54,58 +57,79 @@ export const getAllJobs = async (req, res) => {
   }
 };
 
-export const updateJob = async (req, res) =>{
+export const updateJob = async (req, res) => {
   try {
+
     const jobId = req.params.id;
-    
+
     const job = await Job.findById(jobId);
 
     if (!job) {
       return res.status(404).json({
-        message:"Job not found "
-      })
+        message: "Job not found",
+      });
+    }
+
+    if (
+      req.user.role !== "admin" &&
+      job.recruiter.toString() !== req.user.id
+    ) {
+      return res.status(403).json({
+        message: "You can update only your own jobs",
+      });
     }
 
     const updatedJob = await Job.findByIdAndUpdate(
-      jobId, 
+      jobId,
       req.body,
       {
-        new:true,
+        new: true,
       }
-    )
+    );
 
     res.status(200).json({
-      message:"Job updated successfully",
-      updatedJob
-    })
+      message: "Job updated successfully",
+      updatedJob,
+    });
+
   } catch (error) {
     res.status(500).json({
-      message:error.message
-    })
+      message: error.message,
+    });
   }
 };
 
-export const deleteJob = async (req, res) =>{
+export const deleteJob = async (req, res) => {
   try {
+
     const jobId = req.params.id;
 
     const job = await Job.findById(jobId);
 
     if (!job) {
       return res.status(404).json({
-        message:"Job not found"
+        message: "Job not found",
+      });
+    }
+
+    if (
+      req.user.role !== "admin" &&
+      job.recruiter.toString() !== req.user.id
+    ) {
+      return res.status(403).json({
+        message: "You can delete only your own jobs",
       });
     }
 
     await Job.findByIdAndDelete(jobId);
 
     res.status(200).json({
-      message:"Jon deleted successfully"
-    })
+      message: "Job deleted successfully",
+    });
+
   } catch (error) {
     res.status(500).json({
-      message:error.message
-    })
-    
+      message: error.message,
+    });
   }
-}
+};
